requirements.txt
#ipykernel>=6.29.5
#ipywidgets>=8.1.5
kornia>=0.8.1
matplotlib>=3.10.5
mlflow>=3.4.0
notebook>=7.4.5
numpy>=1.26.4
pandas>=2.3.2
psutil>=7.0.0
pytorch-lightning>=2.5.0
pyyaml>=6.0.2
rasterio>=1.4.3
segmentation-models-pytorch>=0.5.0
timm>=1.0.19
torchgeo>=0.5.2
torchmetrics>=1.6.0
torchvision>=0.19.1
whitebox>=2.3.6
hydra-core
hydra-colorlog
